refactor(plugins): clean createImage
