+++ src/plugins/context.ts
@@ -15,7 +15,10 @@ import YAML from "yaml";
 import zod from "zod";
 
 import Image from "../types/image";
+import { downloadFile } from "../utils/download";
 import * as logger from "../utils/logger";
+import { libraryPath } from "../utils/path";
+import { extensionFromUrl } from "../utils/string";
 
 export const modules = {
   $loader: ora,
@@ -49,6 +52,22 @@ export async function createLocalImage(
 
   logger.log(`Creating image from ${path}`);
   img = new Image(thumbnail ? `${name} (thumbnail)` : name);
+  img.path = path;
+  logger.log(`Created image ${img._id}`);
+
+  return img;
+}
+
+export async function createImage(url: string, name: string, thumbnail?: boolean): Promise<Image> {
+  // if (!isValidUrl(url)) throw new Error(`Invalid URL: ` + url);
+  logger.log(`Creating image from ${url}`);
+  const img = new Image(name);
+  if (thumbnail) {
+    img.name += " (thumbnail)";
+  }
+  const ext = extensionFromUrl(url);
+  const path = libraryPath(`images/${img._id}${ext}`);
+  await downloadFile(url, path);
   img.path = path;
   logger.log(`Created image ${img._id}`);
