+++ src/plugins/events/actor.ts
@@ -9,12 +9,9 @@ import Actor from "../../types/actor";
 import { isValidCountryCode } from "../../types/countries";
 import Image from "../../types/image";
 import Label from "../../types/label";
-import { downloadFile } from "../../utils/download";
 import * as logger from "../../utils/logger";
 import { filterInvalidAliases, validRating } from "../../utils/misc";
-import { libraryPath } from "../../utils/path";
-import { extensionFromUrl } from "../../utils/string";
-import { createLocalImage } from "../context";
+import { createImage, createLocalImage } from "../context";
 
 // This function has side effects
 export async function onActorCreate(
@@ -41,18 +38,8 @@ export async function onActorCreate(
       return img._id;
     },
     $createImage: async (url: string, name: string, thumbnail?: boolean) => {
-      // if (!isValidUrl(url)) throw new Error(`Invalid URL: ` + url);
-      logger.log(`Creating image from ${url}`);
-      const img = new Image(name);
-      if (thumbnail) {
-        img.name += " (thumbnail)";
-      }
-      const ext = extensionFromUrl(url);
-      const path = libraryPath(`images/${img._id}${ext}`);
-      await downloadFile(url, path);
-      img.path = path;
+      const img = await createImage(url, name, thumbnail);
       await Image.setActors(img, [actor._id]);
-      logger.log(`Created image ${img._id}`);
       await imageCollection.upsert(img._id, img);
       if (!thumbnail) {
         createdImages.push(img);
